Added: Foundation for implementing new dictionary parsers
