+++ .gitignore
@@ -1,3 +1,10 @@
 /target
 /data
 /playground
+
+._target
+# idk
+._.DS_Store
+.DS_Store
+
+.vscode/launch.json
