+++ src/importer/parser/base.rs
@@ -1,0 +1,7 @@
+use std::fs::File;
+
+use crate::importer::dictionary::Dictionary;
+
+pub trait DictionaryParser {
+    fn parse(self: &Self, file: File) -> Dictionary;
+}
